@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
-    "react-scripts": "^2.0.5",
+    "react-scripts": "^5.0.1",
     "redux-devtools-extension": "^2.13.5"
   },
   "dependencies": {
@@ -11,12 +11,14 @@
     "@tensorflow/tfjs-backend-webgpu": "^4.22.0",
     "chart.js": "^4.5.1",
     "jwt-decode": "^2.2.0",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
-    "react-redux": "^5.0.7",
-    "react-router-dom": "^4.3.1",
+    "lodash-es": "^4.17.21",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-lifecycles-compat": "^3.0.4",
+    "react-redux": "^8.1.3",
+    "react-router-dom": "^6.30.1",
     "redux": "^4.0.1",
-    "redux-form": "^7.4.2",
+    "redux-form": "^8.3.10",
     "redux-thunk": "^2.3.0"
   },
   "scripts": {
