@@ -1,14 +1,14 @@
 import React from 'react';
 import {connect} from 'react-redux';
-import {Link, Redirect} from 'react-router-dom';
+import {Link, Navigate} from 'react-router-dom';
 
 import RegistrationForm from './registration-form';
 
 export function RegistrationPage(props) {
     // If we are logged in (which happens automatically when registration
-    // is successful) redirect to the user's dashboard
+    // is successful) redirect to the learning page
     if (props.loggedIn) {
-        return <Redirect to="/dashboard" />;
+        return <Navigate to="/learn" replace />;
     }
     return (
         <div className="home">
