@@ -1,6 +1,6 @@
 import React from 'react';
 import {connect} from 'react-redux';
-import {Redirect} from 'react-router-dom';
+import {Navigate} from 'react-router-dom';
 
 export default () => Component => {
     function RequiresLogin(props) {
@@ -8,7 +8,7 @@ export default () => Component => {
         if (authenticating) {
             return <div>Logging in...</div>;
         } else if (!loggedIn || error) {
-            return <Redirect to="/" />;
+            return <Navigate to="/" replace />;
         }
 
         return <Component {...passThroughProps} />;
