@@ -1,5 +1,6 @@
 import React from 'react';
 import {connect} from 'react-redux';
+import {Link} from 'react-router-dom';
 import {clearAuth} from '../actions/auth';
 import {clearAuthToken} from '../local-storage';
 
@@ -12,14 +13,22 @@ export class HeaderBar extends React.Component {
     render() {
         // Only render the greeting and the log out button if we are logged in
         let logOutButton;
+        let navLinks;
         if (this.props.loggedIn) {
             logOutButton = (
                 <button onClick={() => this.logOut()}>Log out</button>
             );
+            navLinks = (
+                <nav className="nav-links">
+                    <Link to="/learn">Learn</Link>
+                    <Link to="/stats">Stats</Link>
+                </nav>
+            );
         }
         return (
             <div className="header-bar">
-                <h1>Speakeasy</h1>
+                <h1>🧠 Neural Learning</h1>
+                {navLinks}
                 {logOutButton}
             </div>
         );
